Rearrange the Git menu some more.
Move things around so that Add, Remove, and Ignore are right above Checkout and Unstage files,
since those operations are opposites of each other. Put Patches and Branches menu, and Create Tag
together on the menu.

I also moved around the actions to coincide with their relative positions on the menu, like I had
before we started messing with the menu. Try to stay this way to keep organized.