symbol-db: db migration updates symbols
If we're in the case of upgrading the db from version 1 to 228 we'll rescan the symbols of the dbs so that the things can be taken consistent.
This should be user invisible, apart of some cpu usage after the first project opening.
Ps: to force the update of symbols run this query on .anjuta_sym_db.db file: 'update version set sdb_version = 1;'