symbol-db: Properly disable the Global tab
Disable the Global tab when only C projects are used instead of just hiding
it. Simply hiding it can cause situations where the buttons and the underlying
notebook are out of sync (i.e. the Global button is active but the Local or
Search pages are displayed.)

Furthermore, explicitly disabling the Global button gives the user
indication that the Search tab should be used instead.