git: Implement the status pane
Basic implementation of the status view. For now there is no select all
option. I'm working on an elegant and user friendly way to to this...

There's also no automatic refreshing for now. This is an even trickier
problem, because there's no fully automated way to get all status changes
in every case, at lest that I know of.