libanjuta: Add a basic stub implementation of AnjutaFileList
This is a new widget I'm building to make building of dialogs that operate on
files, like VCS add and remove dialogs, more uniform and easier to write.
When it's finished, it will allow the user to create a list of files, or file
patterns, to operate on. Users will be able to drag from any source that
supports GFile, or manually enter file paths into the list.