git: Implement automatic ref handling
WARNING: This is completely untested. If it works, it should form the basis of
automatic log refreshing. It should be able to tell us if any kind of ref (branch,
tag, remote, or repository head commit) changes, and the log viewer will,
when it is put together, respond to this command to keep its view updated.