libanjuta: Don't depend on GtkEntryBuffer to catch AnjutaEntry changes.
We can't be guaranteed that the buffer will stay the same throughout the life
of the entry, so we have to use a semi-override of gtk_entry_set_text instead.

This commit introduces a new method, anjuta_entry_set_text that will set the
help mode appropriately, instead of relying on the buffer to notify of changes.