gtodo: Remove todo manager plugin
I mentioned this some time back on the mailing list and on IRC and nobody really complained. We might add something back at some point that uses bugzilla and probably other bug-trackers as backend but for now porting this to GSettings would be half a rewrite and it's not worth it.