language-support-vala: refactor to work with latest project manager
Move everything to the current-editor value handler as the plugin
is activated based on file load rather than project load. Also listen
for the project_loaded signal if the project isn't loaded yet.

This will also allow for more correct behaviour of only parsing vala
files of the target that the currently loaded file belongs to.