Change path of UI file in PyGTK project
A freshly-generated PyGTK project doesn't work out of the box because
the UI file is given without a path. The file is generated in the src/
directory, but the script is executed from the project root directory.
This patch sets UI_FILE='src/whatever.ui' so that a project will run out
of the box.