Alter styling and position of "show all plugins" checkbox
The checkbox to show all Anjuta plugins previously was placed at
the top of the notebook page. This detracted from what the user is
trying to do and can steal users eye focus. Additionally, the amount
of padding allocated for said checkbox stole valuable vertical space
from the plugin list.

This changeset borrows design from the keyboard shortcut list in
the gnome-control-center. It uses a toolbar at the bottom of the
treeview that looks attached to the treeview. The checkbox is placed
within a toolitem in the toolbar.

This saves valuable vertical space and allows the user to more easily
focus on the plugin list.

Signed-off-by: Christian Hergert <chris@dronelabs.com>

https://bugzilla.gnome.org/show_bug.cgi?id=645659