sourceview: Fix a crash with Scintilla
The implementation of ianjuta_document_get_filename in sourceview was returning
a string that the caller has to free while the documentation specify that
it is a constant string.
It causes a crash when using Scintilla due to this recent commit in the
document manager a0571b23499b308cd9bd02ecf71e30a6672947ef.