git: Fix switch delete branch, and merge action sensitivity
The swtich action doesn't care about how many items in the branch pane are
ticked;it switches to the branch that is highlighted, which is usually the last
branch item the user clicked on.

When deleting branches, the check boxes can be checked either before or after
running the action. The selections aren't considered until the OK button the the
dialog is clicked.

Finally, merging doesn't requre any branch to be selected at all, because users
can merge with any git reference, including branches, tags, and even arbitrary
commits.