project-manager: Fix new project manager combo box when the glade plugin is loaded
The new AnjutaPmChooserButton widget is defined in the project manager plugin not in libanjuta.
Anjuta glade library doesn't include this widget and glade creates a dummy widget for it (using parent attribute).
But when the glade plugin is loaded in Anjuta, if it creates this dummy widget, the project manager cannot create the real widget later.
So the project manager plugin must be loaded and create the new widget type before the glade plugin.