git: Use tree paths instead of iterators to track the selected log viewer branch
Tree paths are stored as pointers, so there is no need to make copies of them
when storing them in the branch hash table that the log viewer uses to keep
track of which combo box entry belongs to a branch. Invalid paths can also
easily be denoted with a NULL value, somthing that isn't easily done with
iterators.

This should fix bgo #654094