indentation-c-style: be more picky when completing braces.
We now look at the next character to see if we should autocomplete or not.
The reason for this is that you often want to enclose an already existing
expression in brackets. If we then autocomplete you have to remove this extra
bracket since it's at the wrong place anyways.

https://bugzilla.gnome.org/show_bug.cgi?id=687216