anjuta: Allow to abort loading of a session
This is useful to make the code loading files more general. The old code
was checking which file is a project file because such file triggers the
loading of a new session. In case there was a project session only the project
session was loaded.

The new code load all files in the same way. But if a project file is loaded
it aborts the current session and the project session is loaded instead.