indentation-c-style: Fix get_line_indentation().
get_line_indentation() was broken since 166f7f2cb4eaabb4b148841881cccf2855fa081a. Fix it
by only looking for matching left brace if the current line has a right brace.

Also break on the first line where the number of left braces are more than or equal to the right
braces instead of just breaking on the line where the total number of left and right braces are
equal.

https://bugzilla.gnome.org/show_bug.cgi?id=688275