symbol-db: When updating a symbol try to update the closest one.
This way we get rid of alot of warnings like this:
(lt-anjuta:27226): libanjuta-symbol-db-WARNING **: SQL execute_non_select failed: columns name, file_defined_id, file_position are not unique

which was caused by updating two symbols that has the same name but is at different file positions.
Previously it could happen that we updated the first symbol with the file position of the second
which caused the unique constraint to fail.

https://bugzilla.gnome.org/show_bug.cgi?id=689033