sourceview: Don't store the builder for the preferences in a global variable.
There's no need to keep the builder alive since anjuta_preferences_add_from_builder()
will keep the widgets alive and we can just store all the widgets we want to interact with
in the plugin object.

This fixes a crasher when more than one preferences window is used at the same time for
different windows.

https://bugzilla.gnome.org/show_bug.cgi?id=689054