sourceview: Add marks to an unloaded document the same way we reload marks.
Instead of using a lot of idle sources we can instead use the existing code for reloading marks
to add the new mark when the document is loaded. This fixes a problem where the added
idles would starve all other mainloop sources with a lower priority than
G_PRIORITY_DEFAULT_IDLE.

For example this would cause the loading of documents to completely fail when using GLib
from git since the async read in sourceview-io.c uses G_PRIORITY_LOW. Previous versions
of GLib would complete the asynchronous task using an idle with G_PRIORITY_DEFAULT
regardless of the passed in priority. GLib from git instead completes the task using a idle
with the passed in priority so the idle sources for adding marks now starves the completion of
the asynchronous read.

https://bugzilla.gnome.org/show_bug.cgi?id=691090