file-loader: Fix leak of GtkRecentInfos in the RecentChooserMenu.
The list of GtkRecentInfos in MenuPopulateData would leak if the idle source was removed
using g_source_remove(). Fix this by consolidating all the cleanup code into the
idle_populate_clean_up() function. Also fix so that the cleanup is always run and not
only when populate_id has already been set to 0.