sourceview: use etag to only emit "changed" when file was externally modified
We store the etag of the file when it was last saved/read so that we can
know if a change of the file contents was caused by us or externally.
This way we don't have to remove the file monitor during a save,
and we can also get rid of the five second delay when setting up
the file monitor.

https://bugzilla.gnome.org/show_bug.cgi?id=693547